remove resuming ability, keep resume config but complain if true
