+++ tests/unit/training/test_config.py
@@ -1,0 +1,26 @@
+import pytest
+
+from sae_lens.training.config import LanguageModelSAERunnerConfig
+
+TINYSTORIES_MODEL = "tiny-stories-1M"
+TINYSTORIES_DATASET = "roneneldan/TinyStories"
+
+
+def test_sae_training_runner_config_runs_with_defaults():
+    """
+    Helper to create a mock instance of LanguageModelSAERunnerConfig.
+    """
+    # Create a mock object with the necessary attributes
+    _ = LanguageModelSAERunnerConfig()
+
+    assert True
+
+
+def test_sae_training_runner_config_raises_error_if_resume_true():
+    """
+    Helper to create a mock instance of LanguageModelSAERunnerConfig.
+    """
+    # Create a mock object with the necessary attributes
+    with pytest.raises(ValueError):
+        _ = LanguageModelSAERunnerConfig(resume=True)
+    assert True
